+++ src/test/java/com/example/demo/implement/auth/JwtTokenManagerTest.java
@@ -7,7 +7,7 @@ import java.time.Instant;
 import java.util.Date;
 import java.util.Map;
 
-import com.example.demo.dto.auth.JwtClaims;
+import com.example.demo.implement.token.JwtClaims;
 import com.example.demo.exception.ExceptionType;
 import com.example.demo.exception.JwtCustomException;
 import com.example.demo.factory.TestJwtFactory;
